Update vmax driver
